refactor: move profile modules to v2
