+++ src/app/comment/service.rs
@@ -1,8 +1,10 @@
 use super::model::{Comment, CreateComment, DeleteComment};
 use crate::app::article::model::{Article, FetchBySlugAndAuthorId};
-use crate::app::profile::model::Profile;
-use crate::app::profile::service::{conver_user_to_profile, ConverUserToProfile};
 use crate::app::user::model::User;
+use crate::appv2::features::profile::entities::profile::Profile;
+use crate::appv2::features::profile::usecases::services::{
+    conver_user_to_profile, ConverUserToProfile,
+};
 use crate::error::AppError;
 use diesel::pg::PgConnection;
 use uuid::Uuid;
