+++ src/app/mod.rs
@@ -2,7 +2,6 @@ pub mod article;
 pub mod comment;
 pub mod favorite;
 pub mod follow;
-pub mod profile;
+pub mod healthcheck;
 pub mod tag;
 pub mod user;
-pub mod healthcheck;
