+++ src/appv2/features/mod.rs
@@ -1,0 +1,1 @@
+pub mod profile;
