+++ src/appv2/features/profile/adapters/mod.rs
@@ -1,0 +1,2 @@
+pub mod controllers;
+pub mod presenters;
