+++ src/appv2/features/profile/entities/mod.rs
@@ -1,0 +1,1 @@
+pub mod profile;
