+++ src/appv2/features/profile/mod.rs
@@ -1,0 +1,3 @@
+pub mod adapters;
+pub mod entities;
+pub mod usecases;
