+++ src/appv2/features/profile/usecases/services.rs
@@ -1,4 +1,4 @@
-use super::model::Profile;
+use super::super::entities::profile::Profile;
 use crate::app::user::model::User;
 use crate::error::AppError;
 use diesel::pg::PgConnection;
