+++ src/appv2/mod.rs
@@ -1,1 +1,2 @@
 pub mod drivers;
+pub mod features;
